Reorganize modules related to formatting
