+++ src/ide/completion/struct_constructor.rs
@@ -6,7 +6,7 @@ use cairo_lang_syntax::node::{TypedSyntaxNode, ast};
 use cairo_lang_utils::Upcast;
 use lsp_types::{CompletionItem, CompletionItemKind};
 
-use crate::ide::ty::format_type;
+use crate::ide::format::types::format_type;
 use crate::lang::analysis_context::AnalysisContext;
 use crate::lang::db::AnalysisDatabase;
 use crate::lang::visibility::peek_visible_in_with_edition;
