+++ src/ide/format/documents.rs
@@ -8,7 +8,10 @@ use crate::lang::lsp::LsProtoGroup;
 use crate::state::StateSnapshot;
 
 /// Format a whole document.
-pub fn format(params: DocumentFormattingParams, state: StateSnapshot) -> Option<Vec<TextEdit>> {
+pub fn format_document(
+    params: DocumentFormattingParams,
+    state: StateSnapshot,
+) -> Option<Vec<TextEdit>> {
     let db = &state.db;
     let file_uri = params.text_document.uri;
     let file = db.file_for_url(&file_uri)?;
