+++ src/ide/format/mod.rs
@@ -1,0 +1,2 @@
+pub mod documents;
+pub mod types;
