+++ src/ide/mod.rs
@@ -2,7 +2,7 @@ pub mod analysis_progress;
 pub mod code_actions;
 pub mod code_lens;
 pub mod completion;
-pub mod formatter;
+pub mod format;
 pub mod hover;
 pub mod inlay_hints;
 pub mod introspection;
@@ -10,4 +10,3 @@ pub mod macros;
 mod markdown;
 pub mod navigation;
 pub mod semantic_highlighting;
-pub mod ty;
